for #2567, move SelectItem to optimize module
